refactor: added generic request handler
